[Refactor/#92] domain.throwable error로 rename
