+++ app/src/main/java/com/swyp/firsttodo/data/repositoryimpl/UserRepositoryImpl.kt
@@ -7,9 +7,9 @@ import com.swyp.firsttodo.core.network.util.ApiResponseHandler
 import com.swyp.firsttodo.data.mapper.toModel
 import com.swyp.firsttodo.data.remote.datasource.UserDataSource
 import com.swyp.firsttodo.data.remote.dto.request.user.ProfileRequestDto
+import com.swyp.firsttodo.domain.error.ProfileError
 import com.swyp.firsttodo.domain.model.user.MyInfoModel
 import com.swyp.firsttodo.domain.repository.UserRepository
-import com.swyp.firsttodo.domain.throwable.ProfileError
 import javax.inject.Inject
 
 class UserRepositoryImpl
