+++ app/src/main/java/com/swyp/firsttodo/domain/error/FamilyError.kt
@@ -1,4 +1,4 @@
-package com.swyp.firsttodo.domain.throwable
+package com.swyp.firsttodo.domain.error
 
 sealed class FamilyError(
     override val message: String?,
