+++ app/src/main/java/com/swyp/firsttodo/domain/error/ProfileError.kt
@@ -1,4 +1,4 @@
-package com.swyp.firsttodo.domain.throwable
+package com.swyp.firsttodo.domain.error
 
 sealed class ProfileError(
     override val message: String?,
