+++ app/src/main/java/com/swyp/firsttodo/domain/error/ScheduleError.kt
@@ -1,4 +1,4 @@
-package com.swyp.firsttodo.domain.throwable
+package com.swyp.firsttodo.domain.error
 
 sealed class ScheduleError(
     override val message: String?,
