+++ app/src/main/java/com/swyp/firsttodo/presentation/hamburger/share/ShareViewModel.kt
@@ -6,9 +6,9 @@ import androidx.lifecycle.viewModelScope
 import com.swyp.firsttodo.core.base.Async
 import com.swyp.firsttodo.core.base.BaseViewModel
 import com.swyp.firsttodo.core.network.model.ApiError
+import com.swyp.firsttodo.domain.error.FamilyError
 import com.swyp.firsttodo.domain.model.family.ConnectedFamilyModel
 import com.swyp.firsttodo.domain.repository.FamilyRepository
-import com.swyp.firsttodo.domain.throwable.FamilyError
 import com.swyp.firsttodo.presentation.common.extension.snackbarMsg
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
