+++ app/src/main/java/com/swyp/firsttodo/presentation/todo/TodoViewModel.kt
@@ -10,15 +10,15 @@ import com.swyp.firsttodo.core.base.BaseViewModel
 import com.swyp.firsttodo.core.common.extension.getDataOrNull
 import com.swyp.firsttodo.core.designsystem.theme.LabelColor
 import com.swyp.firsttodo.core.network.model.ApiError
+import com.swyp.firsttodo.domain.error.ScheduleError
+import com.swyp.firsttodo.domain.error.StickerError
+import com.swyp.firsttodo.domain.error.TodoError
 import com.swyp.firsttodo.domain.model.Role
 import com.swyp.firsttodo.domain.model.ScheduleCategory
 import com.swyp.firsttodo.domain.model.todo.TodoCategoryModel
 import com.swyp.firsttodo.domain.repository.ScheduleRepository
 import com.swyp.firsttodo.domain.repository.StickerRepository
 import com.swyp.firsttodo.domain.repository.TodoRepository
-import com.swyp.firsttodo.domain.throwable.ScheduleError
-import com.swyp.firsttodo.domain.throwable.StickerError
-import com.swyp.firsttodo.domain.throwable.TodoError
 import com.swyp.firsttodo.presentation.common.component.DeleteDialogType
 import com.swyp.firsttodo.presentation.common.extension.snackbarMsg
 import com.swyp.firsttodo.presentation.todo.component.ScheduleBottomSheetType
